+++ src/main/java/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/field/index/lucene/LuceneIndexFieldExtension.java
@@ -46,7 +46,7 @@ public class LuceneIndexFieldExtension implements FieldExtension<LuceneIndex> {
         final String name = MoreObjects.firstNonNull(
                 Strings.emptyToNull(annotation.name().trim()), model + '.' + property);
         final OClass clazz = db.getMetadata().getSchema().getClass(model);
-        final OIndex<?> classIndex = clazz.getClassIndex(name);
+        final OIndex classIndex = clazz.getClassIndex(name);
         final OClass.INDEX_TYPE type = OClass.INDEX_TYPE.FULLTEXT;
         if (!descriptor.initialRegistration && classIndex != null) {
             final IndexValidationSupport support = new IndexValidationSupport(classIndex, logger);