+++ src/test/groovy/ru/vyarus/guice/persist/orient/study/UserChangeInTransactionTest.groovy
@@ -22,11 +22,6 @@ class UserChangeInTransactionTest extends AbstractTest {
     @Inject
     PersistentContext<ODatabaseObject> context
 
-    @Override
-    void setupSecurity() {
-        defaultSecurity()
-    }
-
     def "Change user inside transaction"() {
 
         context.doInTransaction({ db ->