+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Orient 3.1 compatibility (updated to 3.1.12)
+
 ### 4.0.0 (2021-07-04)
 * Update to guice 5.0.1
 * (breaking) Update to Orient 3.0.38